Refactor the Tx Details screen, move data to single UiState class
