+++ app/src/main/java/com/tari/android/wallet/ui/screen/settings/allSettings/AllSettingsViewModel.kt
@@ -210,7 +210,7 @@ class AllSettingsViewModel : CommonViewModel() {
             },
             DividerViewHolderItem(),
             SettingsRowViewHolderItem(resourceManager.getString(all_settings_explorer), vector_all_settings_block_explorer_icon) {
-                _openLink.postValue(networkRepository.currentNetwork.blockExplorerUrl.orEmpty())
+                _openLink.postValue(networkRepository.currentNetwork.blockExplorerBaseUrl.orEmpty())
             }.takeIf { networkRepository.currentNetwork.isBlockExplorerAvailable },
             SettingsTitleViewHolderItem(resourceManager.getString(all_settings_advanced_settings_label)),
             SettingsRowViewHolderItem(resourceManager.getString(all_settings_select_theme), vector_all_settings_select_theme_icon) {
