+++ app/src/main/java/com/tari/android/wallet/ui/screen/tx/details/TxDetailsViewModel.kt
@@ -1,8 +1,6 @@
 package com.tari.android.wallet.ui.screen.tx.details
 
 import android.content.Context
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.SavedStateHandle
 import com.tari.android.wallet.R
 import com.tari.android.wallet.R.string.common_are_you_sure
@@ -11,9 +9,7 @@ import com.tari.android.wallet.R.string.tx_details_cancel_dialog_description
 import com.tari.android.wallet.R.string.tx_details_cancel_dialog_not_cancel
 import com.tari.android.wallet.application.walletManager.WalletManager.WalletEvent
 import com.tari.android.wallet.data.contacts.ContactsRepository
-import com.tari.android.wallet.data.contacts.model.ContactDto
 import com.tari.android.wallet.data.contacts.model.splitAlias
-import com.tari.android.wallet.ffi.FFITxCancellationReason
 import com.tari.android.wallet.model.TxStatus.BROADCAST
 import com.tari.android.wallet.model.TxStatus.COINBASE
 import com.tari.android.wallet.model.TxStatus.COINBASE_CONFIRMED
@@ -42,13 +38,14 @@ import com.tari.android.wallet.ui.dialog.modular.modules.button.ButtonModule
 import com.tari.android.wallet.ui.dialog.modular.modules.button.ButtonStyle
 import com.tari.android.wallet.ui.dialog.modular.modules.head.HeadModule
 import com.tari.android.wallet.ui.dialog.modular.modules.input.InputModule
-import com.tari.android.wallet.ui.screen.tx.details.TxDetailsFragment.Companion.TX_EXTRA_KEY
+import com.tari.android.wallet.ui.screen.tx.details.TxDetailsModel.TX_EXTRA_KEY
 import com.tari.android.wallet.util.extension.collectFlow
 import com.tari.android.wallet.util.extension.getOrThrow
 import com.tari.android.wallet.util.extension.launchOnMain
 import com.tari.android.wallet.util.extension.string
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
+import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.update
 import javax.inject.Inject
 
@@ -57,24 +54,23 @@ class TxDetailsViewModel(savedState: SavedStateHandle) : CommonViewModel() {
     @Inject
     lateinit var contactsRepository: ContactsRepository
 
-    val requiredConfirmationCount: Long? = walletManager.walletInstance?.getRequiredConfirmationCount()
-
-    private val _tx = MutableStateFlow(savedState.getOrThrow<Tx>(TX_EXTRA_KEY))
-    val tx = _tx.asStateFlow()
-
-    private val _cancellationReason = MutableLiveData<String>()
-    val cancellationReason: LiveData<String> = _cancellationReason
-
-    private val _explorerLink = MutableLiveData("")
-    val explorerLink: LiveData<String> = _explorerLink
-
-    private val _contact = MutableStateFlow<ContactDto?>(null)
-    val contact = _contact.asStateFlow()
-
     init {
         component.inject(this)
+    }
 
-        collectFlow(contactsRepository.contactList) { updateContact() }
+    private val _uiState = MutableStateFlow(
+        TxDetailsModel.UiState(
+            tx = savedState.getOrThrow<Tx>(TX_EXTRA_KEY),
+            requiredConfirmationCount = walletManager.requireWalletInstance.getRequiredConfirmationCount(),
+            blockExplorerUrl = networkRepository.currentNetwork.blockExplorerBaseUrl,
+        )
+    )
+    val uiState = _uiState.asStateFlow()
+
+    init {
+        collectFlow(contactsRepository.contactList.map { contactsRepository.getContactForTx(this.uiState.value.tx) }) { contact ->
+            _uiState.update { it.copy(contact = contact) }
+        }
 
         collectFlow(walletManager.walletEvent) { event ->
             when (event) {
@@ -93,30 +89,34 @@ class TxDetailsViewModel(savedState: SavedStateHandle) : CommonViewModel() {
     fun addOrEditContact() = showEditNameInputs()
 
     fun openInBlockExplorer() {
-        _openLink.postValue(_explorerLink.value.orEmpty())
+        _openLink.postValue(uiState.value.blockExplorerLink.orEmpty())
     }
 
     fun onTransactionCancel() {
-        if (tx.value is PendingOutboundTx && tx.value.direction == OUTBOUND && tx.value.status == PENDING) {
+        if (uiState.value.tx is PendingOutboundTx && uiState.value.tx.direction == OUTBOUND && uiState.value.tx.status == PENDING) {
             showTxCancelDialog()
         }
     }
 
     fun onAddressDetailsClicked() {
-        val walletAddress = contact.value?.walletAddress ?: return
+        val walletAddress = uiState.value.contact?.walletAddress ?: return
 
         showAddressDetailsDialog(walletAddress)
     }
 
-    private fun setTxArg(tx: Tx) {
-        _tx.update { tx }
-        _cancellationReason.postValue(getCancellationReason(tx))
-        generateExplorerLink(tx)
-        updateContact()
+    fun showTxFeeToolTip() {
+        showSimpleDialog(
+            title = resourceManager.getString(R.string.tx_detail_fee_tooltip_transaction_fee),
+            description = resourceManager.getString(R.string.tx_detail_fee_tooltip_desc),
+        )
+    }
+
+    private fun setTx(tx: Tx) {
+        _uiState.update { it.copy(tx = tx) }
     }
 
     private fun cancelTransaction() {
-        val isCancelled = walletManager.requireWalletInstance.cancelPendingTx(this.tx.value.id)
+        val isCancelled = walletManager.requireWalletInstance.cancelPendingTx(this.uiState.value.tx.id)
         if (!isCancelled) {
             showSimpleDialog(
                 title = resourceManager.getString(R.string.tx_detail_cancellation_error_title),
@@ -137,48 +137,14 @@ class TxDetailsViewModel(savedState: SavedStateHandle) : CommonViewModel() {
         )
     }
 
-    private fun updateContact() {
-        val contact = contactsRepository.getContactForTx(this.tx.value)
-        _contact.update { contact }
-    }
-
-    private fun getCancellationReason(tx: Tx): String {
-        val reason = when ((tx as? CancelledTx)?.cancellationReason) {
-            FFITxCancellationReason.Unknown -> R.string.tx_details_cancellation_reason_unknown
-            FFITxCancellationReason.UserCancelled -> R.string.tx_details_cancellation_reason_user_cancelled
-            FFITxCancellationReason.Timeout -> R.string.tx_details_cancellation_reason_timeout
-            FFITxCancellationReason.DoubleSpend -> R.string.tx_details_cancellation_reason_double_spend
-            FFITxCancellationReason.Orphan -> R.string.tx_details_cancellation_reason_orphan
-            FFITxCancellationReason.TimeLocked -> R.string.tx_details_cancellation_reason_time_locked
-            FFITxCancellationReason.InvalidTransaction -> R.string.tx_details_cancellation_reason_invalid_transaction
-            FFITxCancellationReason.AbandonedCoinbase -> R.string.tx_details_cancellation_reason_abandoned_coinbase
-            else -> null
-        }
-
-        return reason?.let { resourceManager.getString(it) } ?: ""
-    }
-
     private fun updateTxData(tx: Tx) {
-        if (tx.id == this.tx.value.id) {
-            setTxArg(tx)
-        }
-    }
-
-    private fun generateExplorerLink(tx: Tx) {
-        (tx as? CompletedTx)?.txKernel?.let { txKernel ->
-            _explorerLink.postValue(
-                resourceManager.getString(
-                    R.string.explorer_kernel_url,
-                    networkRepository.currentNetwork.blockExplorerUrl.orEmpty(),
-                    txKernel.publicNonce,
-                    txKernel.signature,
-                )
-            )
+        if (tx.id == this.uiState.value.tx.id) {
+            setTx(tx)
         }
     }
 
     private fun showEditNameInputs() {
-        val contact = contact.value!!
+        val contact = uiState.value.contact ?: return
 
         val name = (contact.contactInfo.firstName + " " + contact.contactInfo.lastName).trim()
 
@@ -213,8 +179,8 @@ class TxDetailsViewModel(savedState: SavedStateHandle) : CommonViewModel() {
         launchOnMain {
             val firstName = splitAlias(newName).firstName
             val lastName = splitAlias(newName).lastName
-            val contactDto = contact.value!!
-            _contact.update { contactsRepository.updateContactInfo(contactDto, firstName, lastName, contactDto.yat) }
+            val contactDto = uiState.value.contact!!
+            _uiState.update { it.copy(contact = contactsRepository.updateContactInfo(contactDto, firstName, lastName, contactDto.yat)) }
             hideDialog()
         }
     }
